pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agl-colo-report"
version = "0.1.0"
description = "Create structured colonoscopy reports."
authors = ["Thomas J. Lux <lux_t1@ukw.de>"]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.5"
ffmpeg-python = "^0.2.0"
flake8 = "^3.9.2"
numpy = "^1.21.2"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
